refactor: url합치기 및 코드 리팩토링
